feat: improve user module

+++ src/modules/user/user.controller.ts
@@ -6,10 +6,12 @@ import {
   Body,
   Put,
   Delete,
+  BadRequestException,
 } from '@nestjs/common';
+import uuid from 'uuid';
 
-import { CreateUserDto } from './interfaces/create-user.interface';
-import { UpdatePasswordDto } from './interfaces/update-password.interface';
+import { CreateUserDto } from './dto/create-user.dto';
+import { UpdatePasswordDto } from './dto/update-password.dto';
 import { User } from './interfaces/user.interface';
 import { UserService } from './user.service';
 
@@ -24,6 +26,10 @@ export class UserController {
 
   @Get(':id')
   async findOne(@Param('id') id: string): Promise<User> {
+    if (!uuid.validate(id)) {
+      throw new BadRequestException('userId is invalid (not uuid)');
+    }
+
     return this.userService.findOne(id);
   }
 
@@ -37,11 +43,19 @@ export class UserController {
     @Param('id') @Body() id: string,
     updatePasswordDto: UpdatePasswordDto,
   ): Promise<User> {
+    if (!uuid.validate(id)) {
+      throw new BadRequestException('userId is invalid (not uuid)');
+    }
+
     return this.userService.update(id, updatePasswordDto);
   }
 
   @Delete(':id')
   async delete(@Param('id') id: string): Promise<void> {
+    if (!uuid.validate(id)) {
+      throw new BadRequestException('userId is invalid (not uuid)');
+    }
+
     return this.userService.delete(id);
   }
 }
